+++ deval/__init__.py
@@ -16,7 +16,7 @@
 # DEALINGS IN THE SOFTWARE.
 
 # Define the version of the template module.
-__version__ = "1.3.1"
+__version__ = "1.3.2"
 version_split = __version__.split(".")
 __spec_version__ = (
     (10000 * int(version_split[0]))
